+++ src/main/java/com/stoyankirkov/schooldbtest/repository/MarkRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MarkRepository extends JpaRepository<Mark, Long> {
+public interface MarkRepository extends JpaRepository<Mark, Integer> {
 }
