+++ src/main/java/com/stoyankirkov/schooldbtest/repository/SchoolClassRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SchoolClassRepository extends JpaRepository<SchoolClass, Long> {
+public interface SchoolClassRepository extends JpaRepository<SchoolClass, Integer> {
 }
